+++ influxdb3_cache/src/lib.rs
@@ -1,3 +1,7 @@
 //! Crate holding the various cache implementations used by InfluxDB 3
 
+pub mod last_cache;
 pub mod meta_cache;
+
+#[cfg(test)]
+mod test_helpers;
